Added functionality of RPCs
Main functionality of all RPCs included in the PDF are added, minus logic for markBoard which is currently hard-coded to mark off the number 1 until the server can generate random numbers. Additional changes from the original function descriptions are added and the PDF is updated to reflect changes. Screenshots of a test driver are included below.